==> AIEchDx_demo_notebook/2_output_inception_v3_global_average_pooling_layer.ipynb <==
import numpy as np
from keras.applications.inception_v3 import InceptionV3
from keras.preprocessing import image
from keras.preprocessing.image import ImageDataGenerator
from keras.models import Model
from keras.layers import Dense, GlobalAveragePooling2D, Dropout
from keras import backend as K
from keras.optimizers import SGD
from keras.models import load_model
import os
from PIL import Image
import cv2

# load model

model_name = "inception_v3.h5"
model = load_model(model_name)
model = Model(inputs=model.input, outputs=model.get_layer('global_average_pooling2d_1').output)

# print model

for i, layer in enumerate(model2.layers):
    print(i, layer.name)

# output the values from global average pooling layer

dir_ = "/Data_echo"

fold2 = ["train", "validation" ,"test"]

fold3 = ["ASD","DCM","HP","MI","NORM"]

for f1 in fold2:
    for f2 in fold3:
        path2image = os.path.join(os.path.join(dir_,f1), f2)
        all_images = [x for x in sorted(os.listdir(path2image)) if x[-4:] == '.png']
        SZ = 224
        x_data = np.empty((len(all_images), SZ, SZ, 3), dtype='float32')
        for i, name in enumerate(all_images):
            im = cv2.imread(path2image + name, cv2.IMREAD_COLOR)
            im = cv2.resize(im, dsize=(224, 224), interpolation=cv2.INTER_LANCZOS4)
            im = im / 255.
            x_data[i, ::, ::, ::] = im

        y = model.predict(x_data)
        np.savetxt(f1 + "_" + f2 + "_data.txt", y, delimiter=',')
        with open(f1+"_"+f2+"_name.txt", "w") as f:
            for s in all_images:
                f.write(str(s) +"\n")
        f.close()
